Fix more terminal escape char issues
Need to enclose the color specifier end immediately (with "\[ ... \]")
or else terminal can't find the "line begin". This shows when doing
things like Ctrl-A and rev-search commands in history.

Also, use "\033" for the terminal escape char, since not all shells
(specifically iOS bash?) seem to recognize the "\e" / "\E" shortcut.